[ref] add package and refactoring
