+++ app/build.gradle
@@ -36,7 +36,7 @@ dependencies {
     implementation 'junit:junit:4.12'
 
     implementation "com.android.support:appcompat-v7:28.0.0"
-    //testImplementation 'org.junit.jupiter:junit-jupiter:5.7.1'
+    testImplementation 'org.junit.jupiter:junit-jupiter:5.7.1'
     androidTestImplementation 'com.android.support.test:runner:1.0.2'
     androidTestImplementation 'com.android.support.test.espresso:espresso-core:3.0.2'
