+++ app/src/main/java/com/leetcode_kotlin/DoSome.kt
@@ -1,4 +1,3 @@
 package com.leetcode_kotlin
 
-
 fun main() {}
