+++ app/src/main/java/com/leetcode_kotlin/Task/twoSum/TwoSum/LongestCommonPrefix/DoSomething.java
@@ -1,4 +1,4 @@
-package com.leetcode_kotlin.LongestCommonPrefix;
+package com.leetcode_kotlin.Task.twoSum.TwoSum.LongestCommonPrefix;
 
 
 public class DoSomething {
